pyproject.toml
[project]
name = "dammage-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "pillow>=12.2.0",
    "python-multipart>=0.0.26",
    "ultralytics>=8.4.41",
    "uvicorn[standard]>=0.46.0",
]
